fix:Logout action moved from "Header" to "LogoutButton"
